pyproject.toml
[project]
name = "mcpdoc"
version = "0.1.0"
description = "MCP Documentation Server"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.110.0",
    "mcp>=1.6.0",
    "httpx==0.28.1",
    "pyyaml==6.0.1",
    "markdownify==0.11.6",
    "uvicorn==0.34.0",
]

[project.scripts]
mcpdoc = "mcpdoc.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta" 